[runtime]: tempo: thiserror errors, boxed ismp conversion, prover crate
Follows repo conventions:

- Derive the verifier's Error with thiserror instead of a hand-written
  Display impl, and convert into ismp::error::Error through the AnyHow
  variant rather than Custom(String). The unsupported-state-machine case
  becomes an error variant so every ismp error goes through that one path.
- Split the JSON-RPC prover out of the relayer into tempo-prover, a std-only
  sibling of tempo-verifier, matching modules/consensus/<chain>/{verifier,
  prover}. The live mainnet test moves with it, so it now runs as
  `cargo test -p tempo-prover -- --ignored`.
